+++ src/interface_adapter/calculate_point/CalculatePointController.java
@@ -1,4 +1,4 @@
-package interface_adapter;
+package interface_adapter.calculate_point;
 
 import entity.Question;
 import use_case.calculate_point.CalculatePointInputBoundary;
